Continue to strip out bits of internal model Item
